backend/tests/requirements.txt
# E2E Testing Dependencies

# Testing frameworks
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.1.0
pytest-cov==4.1.0

# Browser automation
selenium==4.15.2
webdriver-manager==4.0.1

# API testing
requests==2.31.0

# Data fixtures
faker==20.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Reporting
pytest-html==4.1.1
pytest-json-report==1.5.0
